refactor: move tagged parameter extraction to a dedicated maven module
