+++ zeebe/engine/pom.xml
@@ -72,6 +72,11 @@
 
     <dependency>
       <groupId>io.camunda</groupId>
+      <artifactId>zeebe-feel-tagged-parameters</artifactId>
+    </dependency>
+
+    <dependency>
+      <groupId>io.camunda</groupId>
       <artifactId>zeebe-dmn</artifactId>
     </dependency>
