+++ zeebe/feel-tagged-parameters/src/main/java/io/camunda/zeebe/feel/tagged/impl/FromAiFunction.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.feel.impl;
+package io.camunda.zeebe.feel.tagged.impl;
 
 import java.util.List;
 import org.camunda.feel.context.JavaFunction;
