+++ zeebe/feel-tagged-parameters/src/main/java/io/camunda/zeebe/feel/tagged/impl/TaggedParameter.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.feel.impl.parameter;
+package io.camunda.zeebe.feel.tagged.impl;
 
 import java.util.Map;
